Update graph codes to merge TC codes
