+++ tslib/react/src/index.ts
@@ -3,15 +3,12 @@ export { plotlyDarkTemplate } from "./components/PlotlyDarkMode"
 export { PlotEdf } from "./components/PlotEdf"
 export type { EdfPlotInfo } from "./components/PlotEdf"
 export { PlotHistory } from "./components/PlotHistory"
-export { PlotTCHistory } from "./components/PlotTCHistory"
 export { PlotImportance } from "./components/PlotImportance"
 export { PlotIntermediateValues } from "./components/PlotIntermediateValues"
 export { PlotSlice } from "./components/PlotSlice"
 export { PlotTimeline } from "./components/PlotTimeline"
-export { PlotTCParallelCoordinate } from "./components/PlotTCParallelCoordinate"
 export { PlotParallelCoordinate } from "./components/PlotParallelCoordinate"
 export { TrialTable } from "./components/TrialTable"
-export { TrialTableTC } from "./components/TrialTableTC"
 export { GraphContainer } from "./components/GraphContainer"
 export { useGraphComponentState } from "./hooks/useGraphComponentState"
 export {
